requirements.txt
Django==4.2.23
djangorestframework==3.16.0
django-cors-headers
rest_framework_simplejwt
djangorestframework-simplejwt
